Server/rasa/requirements.txt
rasa==3.6.2
rasa-sdk==3.6.2
pymongo==4.5.0
python-dotenv==1.0.0 